Attemp to prevent unknown error in prod
